Split stats builtins analysis JSON helpers
